requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
xgboost
Flask
dill
mysql-connector-python

#-e .